Drop font size for smallPhone, move getScaledFont.
